Finished buttons and keybinds, working on maps, print buffer and print regions
Major conglomeration of changes from the past two weeks, including a finalization of the keybinding implementation (relegated to the active state for most keybindings, except "System" keybindings handled by the global game object). Last working on an implementation of maps using JSON style file structures, featuring multiple layers, map and cell dimensions and colors or sprites for cells. The print buffer was part of an experiment to make a scrolling text box and needs to be integrated into the print region. Included as subrepositories atl (Advanced Tiled Loader, although I'm not using or even referencing it) and flux (a really neat tweening library found on the forums). Also included a pure Lua JSON encoder/decoder that I found somewhere and am actually using.